Continue types.ts cleanliness
