+++ src/project/types/book/book-bibliography.ts
@@ -20,11 +20,8 @@ import { binaryPath } from "../../../core/resources.ts";
 import { kBibliography, kCsl } from "../../../config/constants.ts";
 import { Metadata } from "../../../config/types.ts";
 
-import {
-  kProjectRender,
-  ProjectContext,
-  projectOutputDir,
-} from "../../project-shared.ts";
+import { kProjectRender, ProjectContext } from "../../types.ts";
+import { projectOutputDir } from "../../project-shared.ts";
 import { inputTargetIndex, resolveInputTarget } from "../../project-index.ts";
 import { WebsiteProjectOutputFile } from "../website/website.ts";
 import { bookMultiFileHtmlOutputs } from "./book-extension.ts";
