+++ src/project/types/book/book-config.ts
@@ -25,7 +25,7 @@ import {
   SidebarItem,
   SidebarTool,
 } from "../../project-config.ts";
-import { kProjectRender, ProjectConfig } from "../../project-shared.ts";
+import { kProjectRender, ProjectConfig } from "../../types.ts";
 
 import {
   isGithubRepoUrl,
