+++ src/project/types/book/book-crossrefs.ts
@@ -20,7 +20,7 @@ import {
   kOutputFile,
 } from "../../../config/constants.ts";
 import { defaultWriterFormat } from "../../../format/formats.ts";
-import { ProjectContext } from "../../project-shared.ts";
+import { ProjectContext } from "../../types.ts";
 import { projectOutputDir } from "../../project-context.ts";
 import { crossrefIndexForOutputFile } from "../../project-crossrefs.ts";
 import { WebsiteProjectOutputFile } from "../website/website.ts";
