+++ src/project/types/book/book.ts
@@ -38,7 +38,7 @@ import { PandocOptions } from "../../../command/render/types.ts";
 import { pandocMetadataPath } from "../../../command/render/render-shared.ts";
 
 import { ProjectCreate, ProjectType } from "../project-type.ts";
-import { ProjectContext } from "../../project-shared.ts";
+import { ProjectContext } from "../../types.ts";
 import {
   crossrefIndexForOutputFile,
   kCrossrefIndexFile,
