+++ src/project/types/website/website-aliases.ts
@@ -9,7 +9,8 @@ import { dirname, extname, join, relative } from "path/mod.ts";
 
 import { ProjectOutputFile } from "../project-type.ts";
 
-import { ProjectContext, projectOutputDir } from "../../project-shared.ts";
+import { ProjectContext } from "../../types.ts";
+import { projectOutputDir } from "../../project-shared.ts";
 import { renderEjs } from "../../../core/ejs.ts";
 import { resourcePath } from "../../../core/resources.ts";
 import { inputTargetIndex, resolveInputTarget } from "../../project-index.ts";
