+++ src/project/types/website/website-sitemap.ts
@@ -10,7 +10,8 @@ import { join, relative } from "path/mod.ts";
 
 import { ElementInfo, SAXParser } from "xmlp/mod.ts";
 
-import { ProjectContext, projectOutputDir } from "../../project-shared.ts";
+import { ProjectContext } from "../../types.ts";
+import { projectOutputDir } from "../../project-shared.ts";
 import { pathWithForwardSlashes, removeIfExists } from "../../../core/path.ts";
 
 import { renderEjs } from "../../../core/ejs.ts";
