+++ src/project/types/website/website.ts
@@ -13,12 +13,8 @@ import { resourcePath } from "../../../core/resources.ts";
 import { dirAndStem } from "../../../core/path.ts";
 import { isHtmlContent } from "../../../core/mime.ts";
 
-import {
-  kProject404File,
-  ProjectContext,
-  projectOffset,
-  projectOutputDir,
-} from "../../project-shared.ts";
+import { kProject404File, ProjectContext } from "../../types.ts";
+import { projectOffset, projectOutputDir } from "../../project-shared.ts";
 
 import {
   ProjectCreate,
